+++ test/testMotorSpeedRegulator/testMotorSpeedCommander.cpp
@@ -36,7 +36,7 @@ void testSetGoalBehindCurrentPosition()
 
     commander.setGoal( goalPos, 5432 );
 
-    long forwardGoalPosLinear = 1234 + 10775776L / 3000;
+    int32_t forwardGoalPosLinear = 1234 + 10775776L / 3000;
 
     TEST_ASSERT_EQUAL( forwardGoalPosLinear, commander.goalPos.getLinearPosition() );
     TEST_ASSERT_EQUAL( 5432, commander.timeGoalMicros );
@@ -61,8 +61,8 @@ void testComputeTargetSpeed()
     RotationalPositionProvider* posProvider = (RotationalPositionProvider*)1234;
     SpeedRegulatorInterface* speedRegulator = (SpeedRegulatorInterface*)4321;
     MotorSpeedCommander commander(posProvider, speedRegulator);   
-    long timeLeftMicros = 1e6;    
-    long clicksLeft = 3000;
+    int32_t timeLeftMicros = 1e6;    
+    int32_t clicksLeft = 3000;
     TEST_ASSERT_EQUAL( 3000, commander.computeTargetSpeedCPS( timeLeftMicros, clicksLeft ) );
 }
 
@@ -74,8 +74,8 @@ void testComputeTargetSpeedAtTimeZero()
     int16_t maxSpeedCPS = 1234;
     commander.config( maxSpeedCPS );
 
-    long timeLeftMicros = 0;    
-    long clicksLeft = 3000;
+    int32_t timeLeftMicros = 0;    
+    int32_t clicksLeft = 3000;
     TEST_ASSERT_EQUAL( maxSpeedCPS, commander.computeTargetSpeedCPS( timeLeftMicros, clicksLeft ) );
 }
 
@@ -88,8 +88,8 @@ void testComputeTargetSpeedNegativeTimeLeft()
     MotorSpeedCommander commander ( posProvider, speedRegulator );
     commander.config( maxSpeedCPS );
 
-    long timeLeftMicros = -1000;    
-    long clicksLeft = 1000;
+    int32_t timeLeftMicros = -1000;    
+    int32_t clicksLeft = 1000;
     TEST_ASSERT_EQUAL( maxSpeedCPS, commander.computeTargetSpeedCPS( timeLeftMicros, clicksLeft ) );
 }
 
@@ -102,8 +102,8 @@ void testComputeTargetSpeedCapSpeed()
     MotorSpeedCommander commander(posProvider, speedRegulator);
     commander.config( maxSpeedCPS );
 
-    long timeLeftMicros = 1000;    
-    long clicksLeft = 1000;
+    int32_t timeLeftMicros = 1000;    
+    int32_t clicksLeft = 1000;
     TEST_ASSERT_EQUAL( maxSpeedCPS, commander.computeTargetSpeedCPS( timeLeftMicros, clicksLeft ) );
 }
 
@@ -113,8 +113,8 @@ void testRunAlreadyPassedGoal()
     int16_t posGoal = 1234 + 500;
     SpeedRegulatorMock speedRegulator;
     int16_t maxSpeedCPS = 5000;    
-    long timeNow = 4321;
-    long timeGoalMicros = 4321 + 200;
+    int32_t timeNow = 4321;
+    int32_t timeGoalMicros = 4321 + 200;
 
     MotorSpeedCommander commander(&posNow, &speedRegulator);
     commander.config( maxSpeedCPS );    
@@ -135,8 +135,8 @@ void testRunSimple()
     int16_t posGoal = 1234 + 500;
     SpeedRegulatorMock speedRegulator;
     int16_t maxSpeedCPS = 5000;    
-    long timeNow = 4321;
-    long timeGoalMicros = 4321 + 250*1000L;
+    int32_t timeNow = 4321;
+    int32_t timeGoalMicros = 4321 + 250*1000L;
 
     MotorSpeedCommander commander( &posNow, &speedRegulator );
     commander.config( maxSpeedCPS );    
@@ -159,8 +159,8 @@ void testRunOverTime()
     int16_t posGoal = 1234+100;
     SpeedRegulatorMock speedRegulator;
     int16_t maxSpeedCPS = 5000;    
-    long timeNow = 4321;
-    long timeGoalMicros = 4321 - 250;    
+    int32_t timeNow = 4321;
+    int32_t timeGoalMicros = 4321 - 250;    
  
     MotorSpeedCommander commander(&posNow, &speedRegulator);    
     commander.config( maxSpeedCPS );
